execute1: Write SPR RAM on completion
In order to simplify the write data selection logic, we move CFAR to
the even side and LR, CTR and TAR to the odd side.  We arrange for
next_nia to be available as the instruction result for branch
instructions which set LR, and for sc to be written to SRR0, and get
rid of the cia + 4 adder.  This simplifies the choices to either
r.e.write_data or CIA for the even side (CIA for SRR0 and CFAR) and
either r.e.write_data or the SRR1 value for an interrupt on the odd
side.  Furthermore this simplifies the bypassing of write data to SPR
reads.  The debug reads always just read the RAM without bypassing.
The logic for deciding whether to bypass or not on each side is
simplified by assuming that any previous write was done by an
instruction, not an interrupt (because no instruction can be executing
in the cycle immediately following an interrupt) and that if the
previous instruction wrote the SPR RAM and the current instruction is
reading SPR RAM, they are both ALU instructions, so the previous
instruction must be completed by the time the current instruction is
executing.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>