execute1: Use main adder to compute CIA + 4
This eliminates the special-purpose adder that computes CIA + 4 (which
is what gets written to LR for b*l instructions and to SRR0 for sc
instructions).  Instead we arrange for CIA on e_in.read_data1 and 4 on
e_in.read_data2, and use the main adder.  The branch offset for b and
bc instructions, which was on e_in.read_data2, is now on
e_in.read_data3 instead.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>