execute1: Allow ALU and FPU instructions to start with load/stores in progress
Now that we can start execution of an ALU or FPU instruction and then
hold its result and defer its side-effects until any outstanding
load/store instructions have finished, we no longer need to include
the in_progress signal from loadstore1 in the busy output from
execute1.

This also means we can get rid of the 'intr_pending' logic in
execute1.  If an interrupt is pending (e.g. a FPU unavailable
interrupt on a load or store) while a previous load or store is in
progress, the interrupt just sits in r.e.interrupt until it is that
instruction's turn to complete.

Since we can now get instructions being flushed, we need to add a
flush input to the divider so that a division that is in progress can
be aborted.  The multiplier doesn't need a flush signal because it is
fully pipelined.

Because the bus from writeback to execute1 is not just about
interrupts (the itag field carries the tag of the instruction being
completed), we rename it from interrupt_in to wb_in and rename the
'valid' field to 'interrupt'.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>