decode1: Reduce number of single-issue instructions
This reduces the set of instructions marked as single-issue to just
attn and mtspr to certain SPRs (the "slow" SPRs and those in the
load/store unit).  The attn instruction is now decoded fully in
decode1 (previously the minor opcode was decoded in execute1).
The 'sgl_pipe' field in the decode tables is now unused.

The instructions that were previously single-issue are: isync, dcbf,
dcbst, dcbt, dcbtst, eieio, icbi, mfmsr, mtmsr, mtmsrd, sync, tlbsync
and wait.  The synchronization instructions are mostly no-ops anyway
due to the in-order nature of the core, and the cache-management
instructions are unimplemented (except for icbi).  The MSR ops don't
need to be single-issue due to the in-order core and the fact that MSR
updates are effective on the following instruction.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>