multiply: Reduce pipeline depth to 2
This changes the default pipeline depth from 3 to 2.  Reducing the
pipeline depth makes integer multiplication and division and
floating-point multiplication, division and square root faster.
The ECP5 still makes timing at 48MHz with the reduced pipeline depth.
Xilinx-based boards use a different multiplier implementation and so
are not affected by this change.

In order not to increase the number of metavalue warnings reported in
the tests, it turns out to be necessary to avoid putting undefined
values into the pipeline.  Thus the pipeline data is set to zero
unless m_in.valid is 1.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>