core: Start of vector (VMX) / VSX support
This extends the register file to 128 entries and effectively
doubles its width in order to hold VSX registers.  The main
(high) part has 128 entries and holds the GPRs, the FPRs and
the upper half of the VRs.  The low part has 64 entries and
holds the low half of the VSRs, that is, the extensions to the
FPRs and the low half of the VRs.  The gspr_index_t is extended
to 7 bits.  The facility field of the decode tables has new
values VEC and VSX, but no vector or VSX instructions have been
added yet.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>