core: Provide a length encoding for 16-byte transfers
Previously, instructions such as lq had length = is8B, and loadstore
used the fact that the instruction was doubled to determine that there
were actually 16 bytes to be transferred in all.  Instead, we now have
an explicit encoding for a 16-byte transfer.

Loadstore1 now only gives an alignment interrupt on the first instance
of a doubled instruction that requires alignment (lqarx, stqcx., and
some forms of lq).

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>