core: Implement vector loads and stores
This implements the lvx, lvxl, lvebx, lvehx, lvewx, stvx, stvxl,
stvebx, stvehx and stvewx instructions.  The element load instructions
(lvebx, levhx, lvewx) on a cacheable location load data from the
cache into the elements of the destination vector that are not being
addressed, because that is what P9 does.  On a non-cacheable location,
the contents of the non-addressed elements are undefined.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>