core: Extend register file to include VSRs
This adds the extra registers needed to complete the VSX register set.
The gspr_index_t is extended to 8 bits in order to address the 160
GSPRs needed to accommodate the GPRs plus the VSRs.  The addressing is
changed so that bit 5 selects between the high and low halves of each
VSR.  The log format changes slightly to accommodate the additional
register address bit.

The HAS_VEC generic parameter has been replaced by HAS_VECVSX, which
enables both vector (VMX) and VSX facilities when true.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>