core: Add instructions for indirect reading and writing of the register file
This adds two implementation-specific instructions: mfrin (move from
register indirect) and mtrin (move to register indirect).  They access
the register file via a sideband mechanism which is not interlocked
with instruction execution.  mfrin RT,RA reads the register file entry
whose index is in the least-significant 8 bits of RA and puts the
result in RT.  mtrin RS,RA writes the value in RS into the register
file entry indexed by RA.  The opcodes are in the OPF sandbox area,
with mfrin being 0x58000000 and mtrin being 0x58000001.

When the MSB of the 8-bit register index is 1, it selects the low half
of the VSR indexed by the LS 6 bits of the index.  When the MSB is 0,
the LS 7 bits of the index are the gspr_index_t value for the VSR (high
half) or GPR to be accessed.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>