core: Move decoding of prefixed instructions to predecode stage
This extends the predecode logic so that when a prefix is encountered,
it looks at the suffix, and puts the insn_code for the whole prefixed
instruction on the prefix word.  This is to enable us to decode the
suffix differently depending on the prefix word value.  For example,
suffix primary opcode 42 with a 8LS prefix (bit 6 = 0) is plxsd but
with a MLS prefix (bit 6 = 1) is plha.

Then, in decode1, if a prefixed instruction insn_code is encountered,
the insn_code is saved, and when the suffix is encountered, its
insn_code is replaced by the saved insn_code.

If a prefix at an address equal to 60 mod 64 is found, it goes into
the icache as INSN_prefix and then gets the misaligned_prefix flag set
on it in decode1.

All this removes the restriction that the insn_code for a prefixed
instruction has to be one greater than the insn_code for the suffix.
Consequently the insn_code values are reordered slightly and
INSN_op57, INSN_op_60 and INSN_op61 are dropped.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>