core: Implement mfvsr* and mtvsr* instructions
The mfvsr* instructions have an XS source which comes in to execute1
via the C input, so it is most convenient to bring it through the
logical unit - mfvsrd and mfvsrld are like an andc with B=0, and
mfvsrwa is like an extsw but zero-extending rather than
sign-extending.

The mtvsr* instructions have an RA source and possibly an RB source
which come through the A and B inputs, so we bring them through the
misc_result mux.  mtvsrdd and mtvsrws get doubled; the logic in
decode2 that suppresses the A and B inputs to the second instance of
an instruction (to avoid false dependencies on the output of the first
instance) is limited to the DUPD (update-form load) and DRP (load
quadword instructions) cases.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>