core: Implement mfvsr* and mtvsr* instructions
The mfvsrd and mfvsrwz instructions have an XS source which comes
in to execute1 via the C input, so it is most convenient to bring it
through the logical unit - mfvsrd is like an andc with B=0, and
mfvsrwz is like an extsw but zero-extending rather than
sign-extending.

mfvsrld has an XS source, and we want the low doubleword, which is on
e_in.lo_read_data1.  That gets multiplexed through a vector_result
mux which feeds into the main mux.

The mtvsr* instructions have an RA source and possibly an RB source
which come through the A and B inputs, which also come through the
vector_result mux.

We now have a low part to the result which goes to writeback and the
results which go via the bypass paths.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>