predecode: Work out register addressing and instruction class in predecode
This expands the predecode information stored per instruction in the
icache by 12 bits, resulting in 48 bits stored per instruction.  The
extra bits store a 3-bit instruction class and a 9-bit register
addressing field.  The instruction class field indicates whether the
instruction is invalid, a misaligned prefix, a prefixed instruction,
a direct branch (conditional or unconditional), or something else.

The register addressing field is organized as two bits each for the A
and B ports, and four for the C port, and one currently unused bit.
The A and B port bits are encoded as 00 = operand not used, 01 = GPR,
10 = FPR, 11 = VR.  The 4 bits of the C port selector are the top 2
bits of the register address, plus one bit to select which instruction
field, RS or RC, provides the 5 low-order address bits for the C port,
and one bit to force the least significant address bit to one (e.g. for
integer quadword stores).

This greatly simplifies the register address calculations in decode1
and removes constraints on the allocation of insn_code values, at the
expense of some extra block RAM for the icache and for the expanded
predecode tables.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>