core: Implement vector logical instructions
This implements vand, vandc, vnand, vor, vorc, vnor, vxor and veqv.
They are implemented as doubled instructions that use the logical
unit.  The logical unit takes its operands from the C and B register
ports because the integer logical instructions use RS and RB as the
operands.  Hence we provide for the VRA operand to be read via the
C port.

This adds decoding for primary opcode 4 columns 0 - 15 in the
predecode stage.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>