core: Implement VSX logical instructions
This implements xxland, xxlandc, xxlor, xxlxor, xxlnand, xxlnor,
xxlorc and xxleqv, as for the corresponding vector instructions.

We now decode primary opcode 60.  A new decode table of the same size
as the vector decode table has been added.  It contains decodes for 16
column-pairs of 32 rows, so bits 5:2 of the instruction are used to
index the second half compared to 3:0 for the the vector table.  The
columns are decoded in pairs because almost all VSX instructions use
bit 1 as the BX bit.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>