core: Reduce use of instruction word in FPU
Decoding the instruction word in the FPU module makes it difficult to
reuse FPU function for vector or VSX floating-point operations, since
their encoding is usually different from the basic floating-point
instructions.  This uses the sub_select, invert_a and invert_out
fields of the decode table along with the insn_type to control the
selection of the operation to be performed.  Some things still depend
on the instruction word, however, including the selection of the
various mffs* operations, and the bits written to the FPSCR for the
mtfsf* instructions.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>