core: Implement an emulation mode to assist with vector/VSX emulation
This creates an implementation-specific "emulation mode", controlled
by a bit in a new SPR, which has two effects: the GPRs get switched to
a second bank, and CFAR updates are suppressed.  The second GPR bank
uses GSPR numbers 32 - 63, which were previously unused.

The new SPR is number 727, in the sandbox area, and the
least-significant bit is the only bit implemented.  Writes to the SPR
don't take effect until the next event that causes an interrupt or
redirect (isync, rfid, mispredicted branch, etc.).  The emulation mode
bit gets sent to fetch1 and passed down the pipeline, where it affects
register addressing.

In order to avoid having more than 16 slow SPR select values,
SPRSEL_TB and SPRSEL_TBU have been combined into one, and SPRSEL_EMUC
uses the value that SPRSEL_TBU previously had.

The emulation mode of the core gets reported up to the soc and
toplevel layers, and on the Arty A7 and ECPIX-5 causes the relevant
status LED to light the red sub-LED.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>