soc: Add emulation ROM at 0x80000000
This adds 32kB of block RAM to the address map of the SoC, intended to
be used for instruction emulation code.  Although it is called "EROM",
it is currently writable, to assist in development by allowing runtime
updates.  Platforms can opt out of having the emulation ROM by setting
the HAS_EMUL_ROM generic to false, but none do so at present.

The block RAM is initialized from a file called emulation.bin in the
current directory when core_tb is run, or from fpga/emulation.hex when
synthesis is being done.  Thus the test scripts that use core_tb need
to copy emulation.bin to the directory where core_tb is being run.
The emulation.bin and fpga/emulation.hex added here contain 4
instructions which clear emulation mode and jump to the HEAI vector at
0xe40, as follows:

	li	%r0,0
	mtspr	EMUC,%r0
	isync
	ba	0xe40

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>