core: Use emulation ROM to emulate unimplemented instructions
This causes unimplemented instructions to cause an emulation
interrupt, which is like a HEAI (hypervisor emulation assistance
interrupt) but for which the interrupt vector is the start of the
emulation ROM.  The emulation interrupt also turns on emulation mode.
Unimplemented instructions are those that are decoded to OP_ILLEGAL,
but not those that cause a HEAI for other reasons, such as
unimplemented SPR numbers in mfspr/mtspr, or lq with RA = RT.
Emulation interrupts are never taken as if LPCR[HAIL] = 1; they
never leave the MMU enabled or use an alternate interrupt address.

The emulation code in emulation.bin and emulation.hex currently just
turns off emulation mode and jumps to the HEAI interrupt vector at
0xe40, thus preserving the behaviour that unimplemented instructions
cause a HEAI.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>