core: Use emulation ROM to emulate unimplemented vector and VSX instructions
This predecodes unimplemented instructions in primary opcodes 4 and 60
to two new insn_code values, INSN_vemu and INSN_xemu, which then get
decoded to OP_XEMU, which causes an emulation interrupt, which is like
a HEAI (hypervisor emulation assistance interrupt) but for which the
interrupt vector is the start of the emulation ROM.  The emulation
interrupt also turns on emulation mode.

The emulation code in emulation.bin and emulation.hex currently just
turns off emulation mode and jumps to the HEAI interrupt vector at
0xe40, thus preserving the behaviour that unimplemented instructions
cause a HEAI.

The reason for having two separate insn_code values is so that the
primary opcode can be recovered correctly.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>