loadstore1: Make dcbz clear 128 bytes
This makes it easier to run code compiled for POWER9/POWER10.  We use
the DUPD repeat code on dcbz to indicate that it should be doubled
(DUPD is a bit of a misnomer, but currently it's only used on loads,
and using DRP rather than DUPD avoids the illegal form check for
lq/stq/etc in decode2).  The dcbz goes down to loadstore1 twice and
does 64 bytes each time.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>